+++ luminous/routes/listing.js
@@ -4,152 +4,49 @@ const Review = require("../models/review.js");
 const ExpressError = require("../utils/express-error.js");
 const wrapAsync = require("../utils/wrap-async.js");
 const User = require("../models/user.js");
-const route = Router();
+const { onlyLoggedInUser, isLoggedInCheck } = require("../middlewares/auth.js");
+const {
+  handlePostReview,
+  handleDeleteListing,
+  handleCreateLising,
+  handleReadUsernameListing,
+  handleReadListing,
+} = require("../controllers/listing.js");
 
-route.get("/new", (req, res) => {
-  let user = req.user;
-  res.status(200).render("newListing.ejs", { title: "new listing...", user });
-});
+const route = Router();
 
 // async function are wrappred.
 route.get(
   "/",
   wrapAsync(async (req, res) => {
-    let user = req.user;
+    let user = req.user || null;
     let listings = await Listing.find({}).sort({ createdAt: -1 });
-    listings = listings.filter((item) => {
-      return item.createdBy != user._id;
+    res.status(200).render("listings.ejs", {
+      listings,
+      myListings: false,
+      user,
+      title: "listings!!!",
     });
-    res
-      .status(200)
-      .render("listings.ejs", {
-        listings,
-        myListings: false,
-        user,
-        title: "listings!!!",
-      });
   })
 );
+
+route.get("/new", onlyLoggedInUser, (req, res) => {
+  let user = req.user;
+  res.status(200).render("newListing.ejs", { title: "new listing...", user });
+});
 
 route.get(
   "/user/:username",
-  wrapAsync(async (req, res) => {
-    let user = req.user;
-    const { username } = req.params;
-    if (username !== user.username.toString()) {
-      throw new ExpressError(400, "Bad Requiest!! incorrect username!");
-    }
-    const listings = await Listing.find({ createdBy: user._id }).sort({
-      createdAt: -1,
-    });
-    res.status(200).render("listings.ejs", {
-      listings,
-      user,
-      myListings: true,
-      title: "my listings!!",
-    });
-  })
+  onlyLoggedInUser,
+  wrapAsync(handleReadUsernameListing)
 );
 
-route.post(
-  "/new",
-  wrapAsync(async (req, res) => {
-    const { listing } = req.body;
-    let user = req.user;
-    const newListing = new Listing(listing);
-    newListing.createdBy = user._id;
-    await newListing.save();
-    res.status(200).redirect("/listings");
-  })
-);
-
-route.get(
-  "/:id",
-  wrapAsync(async (req, res, next) => {
-    let user = req.user;
-    const { id } = req.params;
-    if (id.toString().length != 24) {
-      throw new ExpressError(400, "Listing id is incorrect!!");
-    }
-    const listing = await Listing.findById(id).populate("reviews");
-    if (listing.createdBy != user._id) {
-      var listingCreatedBy = await User.findById(listing.createdBy);
-    }
-    if (!listing) {
-      // throw new ExpressError(404, "Listing Not Found!!");  // async fuction can throw errors this only if they are wrapped with async_wrapper.
-      throw new ExpressError(404, "Listing not Found!!");
-    }
-    res.status(200).render("listing.ejs", {
-      listing,
-      user,
-      listingCreatedBy,
-      title: "listing based on title",
-    });
-  })
-);
-
-route.post(
-  "/:id/review",
-  wrapAsync(async (req, res) => {
-    let user = req.user;
-    const { id } = req.params;
-    const { rating, msg } = req.body;
-    if (id.toString().length != 24) {
-      throw new ExpressError(400, "Listing id is incorrect!!");
-    }
-    const listing = await Listing.findById(id);
-    if (!listing) {
-      throw new ExpressError(404, "Listing not Found!!");
-    }
-    if (listing.createdBy.toString() === user._id.toString()) {
-      throw new ExpressError(403, "You can't rate your own listing!!");
-    }
-    const popListing = await Listing.findById(id).populate("reviews");
-    const reviews = popListing.reviews.filter((value) => {
-      return value.username === user.username;
-    });
-    if (reviews[0]) {
-      const review = reviews[0];
-      review.rating = rating || review.rating;
-      review.msg = msg || review.msg;
-      await review.save();
-      res.status(201).redirect(`/listings/${listing._id}`);
-    }
-    if (!rating) {
-      throw new ExpressError(400, "Bad Req!! No rating stars provided.");
-    }
-    if (!msg.trim()) {
-      throw new ExpressError(400, "Bad Req!! No msg.");
-    }
-    const review = new Review({
-      rating,
-      msg,
-      username: user.username,
-    });
-    listing.reviews.push(review);
-    await review.save();
-    await listing.save();
-    res.status(201).redirect(`/listings/${listing._id}`);
-  })
-);
-
+route.post("/new", onlyLoggedInUser, wrapAsync(handleCreateLising));
+// unprotected route.
+route.get("/:id", wrapAsync(handleReadListing));
+// post req. for review model
+route.post("/:id", onlyLoggedInUser, wrapAsync(handlePostReview));
 // post route for deleting listing.
-route.post(
-  "/:id/:createdBy",
-  wrapAsync(async (req, res) => {
-    const { id, createdBy } = req.params;
-    if (id.length != 24 || createdBy.length != 24) {
-      throw new ExpressError(400, "Incorrect listing!!");
-    }
-    let user = req.user;
-    // const listing = await Listing.findById(id);
-    if (user._id.toString() !== createdBy) {
-      throw new ExpressError(401, "Unauthorized pruning!!");
-    }
-    await Listing.findByIdAndDelete(id);
-    throw new ExpressError(200, "Listing pruned!!");
-    // res.status(200).redirect("/listing");
-  })
-);
+route.post("/:id/:createdBy", onlyLoggedInUser, wrapAsync(handleDeleteListing));
 
 module.exports = route;
