+++ luminous/views/listings.ejs
@@ -26,7 +26,8 @@
                     &#8377; <%= listing.price.toLocaleString() %>
                   </div>
                   <div>
-                    <%= listing.createdAt.toLocaleString() %>
+                    <%= listing.createdAt.toLocaleDateString("en-IN", { year: 'numeric' , month: 'short' ,
+                      day: 'numeric' , }); %>, <%= listing.createdAt.toLocaleTimeString("th-EG").substring(0, 5) %>
                   </div>
                 </div>
               </div>
